Refactor: 라우팅 이름 변경
